inline enums are not processed
