+++ vertx-web-openapi/src/main/java/io/vertx/ext/web/openapi/RouterBuilderOptions.java
@@ -2,11 +2,8 @@ package io.vertx.ext.web.openapi;
 
 import io.vertx.codegen.annotations.DataObject;
 import io.vertx.codegen.annotations.Fluent;
-import io.vertx.codegen.annotations.VertxGen;
 import io.vertx.core.Handler;
 import io.vertx.core.json.JsonObject;
-
-import java.util.function.Function;
 
 /**
  * @author Francesco Guardiani @slinkydeveloper
@@ -194,23 +191,4 @@ public class RouterBuilderOptions {
     return this;
   }
 
-  /**
-   * The way an OpenAPI operation is transformed into a route name.
-   */
-  @VertxGen
-  public enum RouteNamingStrategy implements Function<Operation, String> {
-    OPERATION_ID(Operation::getOperationId),
-    OPERATION_OPENAPI_PATH(Operation::getOpenAPIPath);
-
-    private final Function<Operation, String> impl;
-
-    RouteNamingStrategy(Function<Operation, String> impl) {
-      this.impl = impl;
-    }
-
-    @Override
-    public String apply(Operation operation) {
-      return impl.apply(operation);
-    }
-  }
 }
